# fix: replace static buffer allocation on growth