Use `slices.Grow` and follow whatever algorithm it uses internally to pick the next size of the buffer.

So the growth might not be exactly what we asked for, but it should be optimal from Go runtime point of view.