@smira 2025-01-31
No real changes with the benchmark: 

```
name                     old time/op    new time/op    delta
Write/defaults-32           152ns ± 3%     153ns ± 5%    ~     (p=0.841 n=5+5)
Write/growing_buffer-32     154ns ± 5%     152ns ± 3%    ~     (p=0.690 n=5+5)
Write/compression-32        425ns ± 3%     423ns ± 2%    ~     (p=1.000 n=5+5)
Read/defaults-32           11.9µs ± 3%    11.9µs ± 2%    ~     (p=0.841 n=5+5)
Read/growing_buffer-32     11.7µs ± 3%    11.9µs ± 2%    ~     (p=0.310 n=5+5)
Read/compression-32         338µs ± 1%     342µs ± 2%    ~     (p=0.151 n=5+5)

name                     old alloc/op   new alloc/op   delta
Write/defaults-32           0.00B          0.00B         ~     (all equal)
Write/growing_buffer-32     0.00B          0.00B         ~     (all equal)
Write/compression-32       1.02kB ± 0%    1.02kB ± 0%    ~     (all equal)
Read/defaults-32             128B ± 0%      128B ± 0%    ~     (all equal)
Read/growing_buffer-32       128B ± 0%      128B ± 0%    ~     (all equal)
Read/compression-32         139kB ± 0%     139kB ± 0%  -0.00%  (p=0.016 n=5+4)

name                     old allocs/op  new allocs/op  delta
Write/defaults-32            0.00           0.00         ~     (all equal)
Write/growing_buffer-32      0.00           0.00         ~     (all equal)
Write/compression-32         0.00           0.00         ~     (all equal)
Read/defaults-32             1.00 ± 0%      1.00 ± 0%    ~     (all equal)
Read/growing_buffer-32       1.00 ± 0%      1.00 ± 0%    ~     (all equal)
Read/compression-32          2.00 ± 0%      2.00 ± 0%    ~     (all equal)
```